Clean up singning
