+++ packages/varsig/src/signing.ts
@@ -1,8 +1,6 @@
 import type { BytesTape } from './bytes-tape.js'
 import { UnreacheableCaseError } from './unreachable-case-error.js'
-import { secp256k1 } from '@noble/curves/secp256k1'
-import * as uint8arrays from 'uint8arrays'
-import { keccak_256 } from '@noble/hashes/sha3'
+import { Secp256k1 } from './signing/secp256k1.js'
 
 type EthAddress = `0x${string}`
 type PublicKey = Uint8Array
@@ -10,11 +8,11 @@ type VerificationKey = PublicKey | EthAddress
 
 export enum SigningKind {
   RSA = 0x1205,
-  SECP256K1 = 0xe7,
+  SECP256K1 = Secp256k1.SIGIL,
 }
 
-export type SigningSecp256k1 = {
-  kind: SigningKind.SECP256K1
+export type SigningAlgo = {
+  kind: SigningKind
   verify: VerifySignatureFn
 }
 
@@ -24,7 +22,6 @@ export type VerifySignatureFn = (
   verificationKey: VerificationKey
 ) => Promise<boolean>
 
-export type SigningAlgo = SigningSecp256k1
 
 export class SigningDecoder {
   constructor(private readonly tape: BytesTape) {}
@@ -37,33 +34,7 @@ export class SigningDecoder {
     const signingSigil = this.tape.readVarint<SigningKind>()
     switch (signingSigil) {
       case SigningKind.SECP256K1: {
-        const recoveryBit = this.tape.read(1)[0]
-        if (recoveryBit && !(recoveryBit === 27 || recoveryBit === 28)) {
-          throw new Error(`Wrong recovery bit`)
-        }
-        return {
-          kind: SigningKind.SECP256K1,
-          verify: async (input, signature, verificationKey) => {
-            let k1Sig = secp256k1.Signature.fromCompact(signature)
-            if (recoveryBit) {
-              k1Sig = k1Sig.addRecoveryBit(recoveryBit - 27)
-              const recoveredKey = k1Sig.recoverPublicKey(input).toRawBytes(false)
-              // compare recoveredKey with verificationKey
-              if (verificationKey instanceof Uint8Array) {
-                return uint8arrays.equals(recoveredKey, verificationKey)
-              }
-              // convert recoveredKey to eth address and compare
-              else if (typeof verificationKey === 'string') {
-                const recoveredAddress = `0x${uint8arrays.toString(
-                  keccak_256(recoveredKey.slice(1)) , 'base16'
-                ).slice(-40)}`
-                return recoveredAddress === verificationKey.toLowerCase()
-              }
-            } else {
-              return secp256k1.verify(signature, input, verificationKey)
-            }
-          }
-        }
+        return Secp256k1.prepareVerifier(this.tape)
       }
       case SigningKind.RSA:
         throw new Error(`Not implemented: signingSigil: RSA`)
@@ -75,7 +46,7 @@ export class SigningDecoder {
   readSignature(signing: SigningAlgo): Uint8Array {
     switch (signing.kind) {
       case SigningKind.SECP256K1: {
-        return this.tape.read(64)
+        return Secp256k1.readSignature(this.tape)
       }
       default:
         throw new UnreacheableCaseError(signing.kind, 'signing kind')
